Rewrote SelfTests to use test specific reporter
The specs in SelfTests.fs were dependent on the original test reporter, and
this the actual output written to the console by this writer, instead of
verifying the behaviour of the test runner.

This rewrite can uses a test specific reporter to capture data sent to the test
reporter in a way the test want to verify behaviour.

Now we can soon delete the old test reporter.