Changed matcher continuation to include actual value
This allows different matchers to modify what is considered the actual
value. E.g. if the matcher input is a function, and the matcher expects an
exception with a specific message, then the matcher failure message should
display the actual exception message, and not the matcher input, which is
the function expected to throw an exception.