Created hardcoded testsuite name
The testsuite abstraction does not have any particular meaning in FSpec -
in particular when the format does not support nested test suites.

It could correspond to a single .dll containing tests, but then that
should be wrapped in a testsuites element.

As a first implementation, it is much simpler to just place all tests in a
single testsuite element, regardless of how many test assemblies are
loaded by the runner.